tests/requirements.txt
pytest==6.0.0
pyyaml
regex
pytest-dependency
